+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # minecraft-server cookbook changelog
 
+## Version 0.1.1 (9/9/2016)
+
+* Add support for installing Bukkit plugins from [bukkit.org](https://www.bukkit.org/)
+* Install plugins with links
+
 ## Version 0.1.0 (9/4/2016)
 
 * Add support for vanilla Minecraft servers